requirements.txt
discord.py
python-dotenv
requests
hentai
genshinstats
cachetools